Update Ex 3 project files
