+++ Demos/03-mockdata/config/config.json
@@ -2,17 +2,17 @@
   "$schema": "https://developer.microsoft.com/json-schemas/spfx-build/config.2.0.schema.json",
   "version": "2.0",
   "bundles": {
-    "sp-fx-content-web-part": {
+    "sp-fx-http-client-demo-web-part": {
       "components": [
         {
-          "entrypoint": "./lib/webparts/spFxContent/SpFxContentWebPart.js",
-          "manifest": "./src/webparts/spFxContent/SpFxContentWebPart.manifest.json"
+          "entrypoint": "./lib/webparts/spFxHttpClientDemo/SpFxHttpClientDemoWebPart.js",
+          "manifest": "./src/webparts/spFxHttpClientDemo/SpFxHttpClientDemoWebPart.manifest.json"
         }
       ]
     }
   },
   "externals": {},
   "localizedResources": {
-    "SpFxContentWebPartStrings": "lib/webparts/spFxContent/loc/{locale}.js"
+    "SpFxHttpClientDemoWebPartStrings": "lib/webparts/spFxHttpClientDemo/loc/{locale}.js"
   }
 }
