+++ Demos/03-mockdata/config/deploy-azure-storage.json
@@ -2,6 +2,6 @@
   "$schema": "https://developer.microsoft.com/json-schemas/spfx-build/deploy-azure-storage.schema.json",
   "workingDir": "./temp/deploy/",
   "account": "<!-- STORAGE ACCOUNT NAME -->",
-  "container": "sp-fx-content",
+  "container": "sp-fx-http-client-demo",
   "accessKey": "<!-- ACCESS KEY -->"
 }
