+++ Demos/03-mockdata/src/models/ButtonClickedCallback.ts
@@ -1,1 +1,4 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 export type ButtonClickedCallback = () => void;
