+++ Demos/03-mockdata/src/models/ICountryListItem.ts
@@ -1,4 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 export interface ICountryListItem {
-  Id: string;
-  Title: string;
+    Id: string;
+    Title: string;
 }
