+++ Demos/03-mockdata/src/models/index.ts
@@ -1,2 +1,5 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 export * from './ButtonClickedCallback';
 export * from './ICountryListItem';
