--- Demos/03-mockdata/src/webparts/spFxContent/SpFxContentWebPart.manifest.json
@@ -1,26 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/json-schemas/spfx/client-side-web-part-manifest.schema.json",
-  "id": "b8356080-c627-400f-ba73-a3e388040f33",
-  "alias": "SpFxContentWebPart",
-  "componentType": "WebPart",
-
-  // The "*" signifies that the version should be taken from the package.json
-  "version": "*",
-  "manifestVersion": 2,
-
-  // If true, the component can only be installed on sites where Custom Script is allowed.
-  // Components that allow authors to embed arbitrary script code should set this to true.
-  // https://support.office.com/en-us/article/Turn-scripting-capabilities-on-or-off-1f2c515f-5d7e-448a-9fd7-835da935584f
-  "requiresCustomScript": false,
-
-  "preconfiguredEntries": [{
-    "groupId": "5c03119e-3074-46fd-976b-c60198311f70", // Other
-    "group": { "default": "Other" },
-    "title": { "default": "SPFxContent" },
-    "description": { "default": "SPFxContent description" },
-    "officeFabricIconFontName": "Page",
-    "properties": {
-      "description": "SPFxContent"
-    }
-  }]
-}
