--- Demos/03-mockdata/src/webparts/spFxContent/SpFxContentWebPart.ts
@@ -1,225 +1,0 @@
-import * as React from 'react';
-import * as ReactDom from 'react-dom';
-import { Version } from '@microsoft/sp-core-library';
-import {
-  BaseClientSideWebPart,
-  IPropertyPaneConfiguration,
-  PropertyPaneTextField
-} from '@microsoft/sp-webpart-base';
-
-import * as strings from 'SpFxContentWebPartStrings';
-import SpFxContent from './components/SpFxContent';
-import { ISpFxContentProps } from './components/ISpFxContentProps';
-
-import { Environment, EnvironmentType } from '@microsoft/sp-core-library';
-import { SPHttpClient, SPHttpClientResponse } from '@microsoft/sp-http';
-import { ICountryListItem } from '../../models';
-
-export interface ISpFxContentWebPartProps {
-  description: string;
-}
-
-export default class SpFxContentWebPart extends BaseClientSideWebPart<ISpFxContentWebPartProps> {
-  private _countries: ICountryListItem[] = [];
-
-  public render(): void {
-    const element: React.ReactElement<ISpFxContentProps > = React.createElement(
-      SpFxContent,
-      {
-        spListItems: this._countries,
-        onGetListItems: this._onGetListItems,
-        onAddListItem: this._onAddListItem,
-        onUpdateListItem: this._onUpdateListItem,
-        onDeleteListItem: this._onDeleteListItem
-      }
-    );
-
-    ReactDom.render(element, this.domElement);
-  }
-  
-  private get _isSharePoint(): boolean {
-    return (Environment.type === EnvironmentType.SharePoint || Environment.type === EnvironmentType.ClassicSharePoint);
-  }
-
-  private _onGetListItems = (): void => {
-    if (!this._isSharePoint) {
-      this._countries = [
-        { Id: '1', Title: 'Country 1' },
-        { Id: '2', Title: 'Country 2' },
-        { Id: '3', Title: 'Country 3' },
-        { Id: '4', Title: 'Country 4' }
-      ];
-      this.render();
-    } else {
-      this._getListItems()
-        .then(response => {
-          this._countries = response;
-          this.render();
-        });
-    }
-  }
-
-  private _onAddListItem = (): void => {
-    if (!this._isSharePoint) { return; }
-
-    this._addListItem()
-      .then(() => {
-        this._getListItems()
-          .then(response => {
-            this._countries = response;
-            this.render();
-          });
-      });
-  }
-
-  private _onUpdateListItem = (): void => {
-    if (!this._isSharePoint) { return; }
-
-    this._updateListItem()
-      .then(() => {
-        this._getListItems()
-          .then(response => {
-            this._countries = response;
-            this.render();
-          });
-      });
-  }
-
-  private _onDeleteListItem = (): void => {
-    if (!this._isSharePoint) { return; }
-
-    this._deleteListItem()
-      .then(() => {
-        this._getListItems()
-          .then(response => {
-            this._countries = response;
-            this.render();
-          });
-      });
-  }
-
-  private _getListItems(): Promise<ICountryListItem[]> {
-    return this.context.spHttpClient.get(
-        this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')/items?$select=Id,Title`, 
-        SPHttpClient.configurations.v1)
-      .then(response => {
-        return response.json();
-      })
-      .then(jsonResponse => {
-        return jsonResponse.value;
-      }) as Promise<ICountryListItem[]>;
-  }
-
-  private _getItemEntityType(): Promise<string> {
-    return this.context.spHttpClient.get(
-        this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')?$select=ListItemEntityTypeFullName`,
-        SPHttpClient.configurations.v1)
-      .then(response => {
-        return response.json();
-      })
-      .then(jsonResponse => {
-        return jsonResponse.ListItemEntityTypeFullName;
-      }) as Promise<string>;
-  }
-  private _addListItem(): Promise<SPHttpClientResponse> {
-    return this._getItemEntityType()
-      .then(spEntityType => {
-        const request: any = {};
-        request.body = JSON.stringify({
-          Title: new Date().toUTCString(),
-          '@odata.type': spEntityType
-        });
-
-        return this.context.spHttpClient.post(
-          this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')/items`,
-          SPHttpClient.configurations.v1,
-          request);
-        }
-      ) ;
-  }
-
-  private _updateListItem(): Promise<SPHttpClientResponse> {
-    // get the first item
-    return this.context.spHttpClient.get(
-        this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')/items?$select=Id,Title&$filter=Title eq 'United States'`, 
-        SPHttpClient.configurations.v1)
-      .then(response => {
-        return response.json();
-      })
-      .then(jsonResponse => {
-        return jsonResponse.value[0];
-      })
-      .then((listItem: ICountryListItem) => {
-        // update item
-        listItem.Title = 'USA';
-        // save it
-        const request: any = {};
-        request.headers = {
-          'X-HTTP-Method': 'MERGE',
-          'IF-MATCH': (listItem as any)['@odata.etag']
-        };
-        request.body = JSON.stringify(listItem);
-
-        return this.context.spHttpClient.post(
-          this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')/items(${listItem.Id})`,
-          SPHttpClient.configurations.v1,
-          request);
-      });
-  }
-
-  private _deleteListItem(): Promise<SPHttpClientResponse> {
-    // get the last item
-    return this.context.spHttpClient.get(
-        this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')/items?$select=Id,Title&$orderby=ID desc&$top=1`, 
-        SPHttpClient.configurations.v1)
-      .then(response => {
-        return response.json();
-      })
-      .then(jsonResponse => {
-        return jsonResponse.value[0];
-      })
-      .then((listItem: ICountryListItem) => {
-        const request: any = {};
-        request.headers = {
-          'X-HTTP-Method': 'DELETE',
-          'IF-MATCH': '*'
-        };
-        request.body = JSON.stringify(listItem);
-
-        return this.context.spHttpClient.post(
-          this.context.pageContext.web.absoluteUrl + `/_api/web/lists/getbytitle('Countries')/items(${listItem.Id})`,
-          SPHttpClient.configurations.v1,
-          request);
-      });
-  }
-
-  protected onDispose(): void {
-    ReactDom.unmountComponentAtNode(this.domElement);
-  }
-
-  protected get dataVersion(): Version {
-    return Version.parse('1.0');
-  }
-
-  protected getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
-    return {
-      pages: [
-        {
-          header: {
-            description: strings.PropertyPaneDescription
-          },
-          groups: [
-            {
-              groupName: strings.BasicGroupName,
-              groupFields: [
-                PropertyPaneTextField('description', {
-                  label: strings.DescriptionFieldLabel
-                })
-              ]
-            }
-          ]
-        }
-      ]
-    };
-  }
-}
