--- Demos/03-mockdata/src/webparts/spFxContent/components/ISpFxContentProps.ts
@@ -1,12 +1,0 @@
-import { 
-  ButtonClickedCallback,
-  ICountryListItem 
-} from '../../../models';
-
-export interface ISpFxContentProps {
-  spListItems: ICountryListItem[];
-  onGetListItems?: ButtonClickedCallback;
-  onAddListItem?: ButtonClickedCallback;
-  onUpdateListItem?: ButtonClickedCallback;
-  onDeleteListItem?: ButtonClickedCallback;
-}
