--- Demos/03-mockdata/src/webparts/spFxContent/loc/en-us.js
@@ -1,7 +1,0 @@
-define([], function() {
-  return {
-    "PropertyPaneDescription": "Description",
-    "BasicGroupName": "Group Name",
-    "DescriptionFieldLabel": "Description Field"
-  }
-});
