--- Demos/03-mockdata/src/webparts/spFxContent/loc/mystrings.d.ts
@@ -1,10 +1,0 @@
-declare interface ISpFxContentWebPartStrings {
-  PropertyPaneDescription: string;
-  BasicGroupName: string;
-  DescriptionFieldLabel: string;
-}
-
-declare module 'SpFxContentWebPartStrings' {
-  const strings: ISpFxContentWebPartStrings;
-  export = strings;
-}
