+++ Demos/03-mockdata/src/webparts/spFxHttpClientDemo/SpFxHttpClientDemoWebPart.manifest.json
@@ -1,0 +1,26 @@
+{
+  "$schema": "https://developer.microsoft.com/json-schemas/spfx/client-side-web-part-manifest.schema.json",
+  "id": "48a92d71-f878-4722-8eb3-c4c5802bcda7",
+  "alias": "SpFxHttpClientDemoWebPart",
+  "componentType": "WebPart",
+
+  // The "*" signifies that the version should be taken from the package.json
+  "version": "*",
+  "manifestVersion": 2,
+
+  // If true, the component can only be installed on sites where Custom Script is allowed.
+  // Components that allow authors to embed arbitrary script code should set this to true.
+  // https://support.office.com/en-us/article/Turn-scripting-capabilities-on-or-off-1f2c515f-5d7e-448a-9fd7-835da935584f
+  "requiresCustomScript": false,
+
+  "preconfiguredEntries": [{
+    "groupId": "5c03119e-3074-46fd-976b-c60198311f70", // Other
+    "group": { "default": "Other" },
+    "title": { "default": "SPFxHttpClientDemo" },
+    "description": { "default": "SPFxHttpClientDemo description" },
+    "officeFabricIconFontName": "Page",
+    "properties": {
+      "description": "SPFxHttpClientDemo"
+    }
+  }]
+}
