+++ Demos/03-mockdata/src/webparts/spFxHttpClientDemo/loc/en-us.js
@@ -1,0 +1,7 @@
+define([], function() {
+  return {
+    "PropertyPaneDescription": "Description",
+    "BasicGroupName": "Group Name",
+    "DescriptionFieldLabel": "Description Field"
+  }
+});
