+++ Demos/03-mockdata/src/webparts/spFxHttpClientDemo/loc/mystrings.d.ts
@@ -1,0 +1,10 @@
+declare interface ISpFxHttpClientDemoWebPartStrings {
+  PropertyPaneDescription: string;
+  BasicGroupName: string;
+  DescriptionFieldLabel: string;
+}
+
+declare module 'SpFxHttpClientDemoWebPartStrings' {
+  const strings: ISpFxHttpClientDemoWebPartStrings;
+  export = strings;
+}
